Feat: Welcome with tag screen and fix search term
Add a notmuch-hello type screen with all the valid tags. Hit enter on
one of them and it will search all threads with that tag.

Fix the name of the search_tag to search_terms, such that we can pass
any notmuch query through the command, just like you would in cli
notmuch queries. For example:

  NotmuchSearch tag:gmail and date:yesterday
                ~~~~~~~~~~~~~~~~~~~~~~~~~~~~

Will perform that notmuch query and return all the output threads to a
scratch buffer. To reiterate, hitting enter on a thread and it will open
the thread view in a floating scratch buffer (see next commit for
compartmentalizaiotn of floating window functions.

Note: We probably don't want to open threads in a floating buffer. Think
about how to open them in (vert) splits based on dimensions of the
current window.