Fix: Well behaved completion for search function
Fix a previous commit (4c59cd9) related to completion function with
respect to the completion behavior. Previously it would return a solid
list of completion items that were static and would not respond to what
the user already has written. For example:

  :NotmuchSearch tag:u<Tab>

Returns the full list of tags, when it should really return the tags
that start with 'u'. This commit fixes this behavior.

Keep in mind that this meant rewriting it in vimscript as opposed to
Lua. This is due to easy access to already provided functions like
`split()` and `join()`. This was a compromise I'm willing to take.
Vimscript is just more suited for completion functions for commands.