WIP: Module for parsing and displaying messages
Started out making a `msg.lua` module for parsing information from
individual files. Ideally, a `parse_thread()` would be looking through
for all the messages in the `notmuch-show` style output.

Each message gets passed into this module to produce meta information
such as pointers to headers, table of body parts, and things like that
which will help us atomize the message and do things on it.