Ftplugin: Wipeout buffers instead of killing
Ensure that when a notmuch buffer is closed, that it is REALLY closed.
With the previous `:bd`, it would still be listed in `:ls!`. See `:h
bwipeout` for more details.

Also add the same thing form `mail.vim` so that when we open messages,
we can easily quit out of it back to the threads view.