Feat: Ability to change command for maildir sync
Maintain a global variable for syncing the maildir, initialized in
'init.lua' (ie. whenever the first Notmuch buffer is created). It can be
customized in a `init.vim` like so:

    let g:NotmuchMaildirSyncCmd = "<Some syncing command>"

This command does not include the `notmuch new` portion of
synchronizing. That is done automatically when a user presses the
binding ('%') to sync.