Feat: Extract and open urls from current msg
Use a custom script called `urlextract` to find all the URLs in the
current message under the cursor, display them in a FZF terminal buffer
(note the fzf dependency) and, when selected, opens the URL in a
browser.

TODO: add dependency guards to check for FZF and urlextract
compatibility. Maybe even keep the latter in a scritps/ dir.