Feat: Follow and display github patches for PRs
Use curl to follow GitHub `.patch` links in pull requests and display
the patch list in a scratch buffer.

This is a vanity feature which is made for convenience of working in
project codebases. The idea is to pull closer the workflow between mail,
GitHub patches, and the codebase text you're editing/reviewing.