Add LuaJIT/C bindings for NeoVim and Notmuch
Add a LuaJIT module which interfaces with the Notmuch C header bindings.
The aim is to make the Notmuch queries and interactions in NeoVim faster
and more focused on library calls rather than spawning subprocesses (the
former is obviously less expensive and takes less overhead).

Some sample use cases in the current plugin setup:

* The `notmuch-hello` shows a list of all tags from bindings
* When searching using `:NmSearch`, access the database to see how many
  threads match the search term.

Also did some plumbing and reduced some of the codebase. Namely, replace
the `show_tag()` function with just a call to `search_terms()` with the
tag as the search term. The result is no different at all in terms of
behavior and performance.