Syntax: Conceal threadid for threads in index
Hide the `thread:xxxxx` information about a thread in the index view as
it is not needed by the user when browsing the index.

It technically is still in the line, which is useful for passing queries
to show a thread or tag it, etc. But it won't bother the user with noise
and long lines anymore.