WIP: Add cnotmuch method to tag messages
Re-implement `msg_add_tag()` in `tag.lua` using the cnotmuch bindings.
Essentially, rewrite the function to add the tag using call to the C
api, as opposed to spawning a child shell process and run notmuch on
that.

Will obviously have to rewrite `msg_tag_rm()` and `msg_tag_toggle()` and
their thread counterparts. Keep in mind that especially for toggling
tags, using the C api is a much more logical solution.