Fix: Notmuch bindings sync flags with maildir
Apply a refresh to the maildir flags when adding or removing tags from
threads/messages in notmuch bindings.

A common case where this was a problem: Remove a 'unread' from a
thread/msg.

Issue: Removes tag but doesn't apply the ',S' to the filename.
After: Synchronizes filename IMAP/Maildir flags with nm tags.