Feat: Customize path to notmuch database
Make the path to a user's notmuch database customizable via a global vim
variable (see `:h global-variable`).

Just like the mail synchronization command, this has a default value
'$HOME/Mail'. This patch allows the user to change this. This is
especially useful for users of different platforms who may have
different file system structures and hierarchies.

For Windows users, as long as you have a set $HOME environment variable,
this shouldn't be a problem.