WIP: First try getting async search
Add asynchronous, non-blocking searching in the notmuch database.

This makes use of the `luvit` library in lua and vim.loop to spawn a
process that searches the database. Instead of blocking the NeoVim main
event loop, you can continue using the editor as the results are
returning.

As the results pour in, a standard output pipe captures the threads that
match the search terms. An `onread()` hook decides what to do with
output data. Currently, it just adds each line to a table.

Moving forward, the goal is to use the `onread()` function to append the
output results into the notmuch buffer as results are coming in,
effectively making a complete asynchronous replacement for the current
searching mechanism.

Credit and References:
- `:help vim.loop`
- https://github.com/luvit/luv/blob/master/docs.md
- https://teukka.tech/vimloop.html