Feat: Customize open command for attachments
Allow user to customize which command to use when generically
opening/viewing attachment parts from emails. This follows commit
548ffa7 in allowing more customizability and portability between
platforms and operating systems.

By default the plugin will attempt to set the open command to open(1) or
xdg-open(1) based on whether you are using a Mac or other operating
system, respectively.

If the user has set `g:NotmuchOpenCmd`, then that command will be used
instead as the open command.