Fix: Customize open command for WSL
Set generic open command introduced from 3ea9606 to `wsl-open(1)` [1] as
opposed to the catch-all linux `xdg-open(1)`. The former is more suited
for opening a larger variety of attachments on WSL. Namely, opening HTML
parts behaves and responds very well as opposed to the latter command,
which acts slow and buggy.