Revert "Remove redundant completion function"
This reverts commit 28da21b80ded07cd294ca7864f2e923c774cb302.

It turns out we do actually need the `notmuch#CompTags()` completion
function for something and it was not redundant after all. Namely, it is
used for commands that take _just_ a tag (or list of tags) as an
argument.

For example, say you want to add the (inbox, unread, flagged) tags to a
thread under the cursor. Then run `:TagAdd ^I`, and it will complete a
list of tags for you. Makes sense.