Fix: Prevent extra buf for existing attachment buf
Prevent an extra, empty, buffer being created when you launch the
attachment list on a message whose attachment list already exists.

Instead of creating a buffer, we check that it exists, and if it does,
return the buffer number to the user. Otherwise, if the attachment list
does not exist, proceed as usual.