Feat: Add `count_messages` in cnotmuch.lua
Add the ability to count the number of messages that match for a given
`query`. Previously, we had a `count_threads()` function which returns a
count of threads that match the query. Now this is extended to count
messages.

Actually, counting messages is the default behavior when invoking the
command line `notmuch count`. Counting by threads turns out to be mega
slow, but messages count fast.

This is especially useful for my plan to add counts next to tags in the
`notmuch-hello` screen.