Feat: Highlight tags in mail.vim
Add the ability to highlight select tags from the mail view message
headline.

This is done using the `mail.vim` syntax file, which is essentially read
by NeoVim when opening a thread. For each message in that thread, there
is a headline for the message that contains a list of tags pertinent to
that particular message.

So far, I have set up highlighting for `inbox`, `unread`, and `flagged`
tags and have tested them. One problem so far is that the highlighting
may extend to other lines in the email OTHER than the headline. I would
need to have better regex-fu to improve/fix this.

The file is split into sections to make it easy to read and change the
list of tags to highlight in the future. Resources to learn more about
the highlighting process in vim are also referenced in the file.