Feat: Add notmuch-count function for notmuch-hello
Add a function `nm.count()` whose behavior corresponds to that of the
`notmuch-count` subcommand in notmuch. It takes a search query string as
an argument (see `man notmuch-search-terms`) and counts and returns the
number of resulting messages from that search query.

First implementation includes binding `c` in `notmuch-hello` buffer to
count messages for the tag name under the cursor's current line. The
result will simply echo the result number of messages in NeoVim's
command line at the bottom. Pretty convenient to count a tag before
choosing to open it.